requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-dotenv
passlib[argon2]
pyjwt
python-multipart
google-api-python-client
google-auth
google-auth-oauthlib